Move interceptor to its own folder.
